+++ NEWS.md
@@ -1,5 +1,14 @@
-# SpiGesR 0.0.3
+# SpiGesXML 0.0.6
+- add format validation message using spiges_get_df()
+- access IDs data with "Personenidentifikatoren" node
+- custom validation format with `schema_xsd` argument
+- improve documentation
+- add unit tests
+- returns error message with suggestions if node or variable doesn't exist
+- return empty tibble with message if no data in node.
+
+# SpiGesXML 0.0.3
 - updated to SpiGes version 4.1
 
-# SpiGesR 0.0.2
+# SpiGesXML 0.0.2
 - fixed "Operierende" node by adding one level down.